+++ Changes.md
@@ -72,6 +72,10 @@ API
   - Added `curve.removeInactiveKeys()` function, removes all the inactive keys parented to a curve.
   - Added `key.isActive()` function, returns true if key is parented to curve and is the active key at its current time.
   - Added `curve.insertKey()` functions, insert key at specified time with optionally specified value.
+  - Added `curve.keyTimeChangedSignal()` function, returns a signal that is called when a key's time has changed
+  - Added `curve.keyValueChangedSignal()` function, returns a signal that is called when a key's value has changed
+  - Added `curve.keyInterpolationChangedSignal()` function, returns a signal that is called when a key's interpolation has changed
+
 - AnimationGadget :
   - Added `selectedKeys()` function, returns current set of selected keys.
   - Added `onTimeAxis()` function, returns true if specified line is over the time axis of the gadget.